add coreutils to docker-support/Dockerfile
* tail --lines=+2 in hack/import-flux2-assets.sh

does not work with busybox tail

* Run "make docker-push-support"
  to rebuild from docker-support/ and push a new image to the repo

(This change will require the maintainer to run this command,
to update the fluxcd/website docker support container images.)

* refactor docker build

It should be a bit easier for the maintainer to push a change to the
docker-support image (you should basically never need to rebuild hugo,
which unfortunately takes 5-10 minutes even on a fast machine.)

This is not a big deal but if someone finds a problem with the docker
support image and they can't actually test and contribute a simple fix
like adding the coreutils package without spending 10 mins needlessly
building the same version of hugo we already are using, that's not a
very good contrib-experience.

* add support for remote builder

Especially when on a slow or remote network, or if your workstation is
underpowered, it can be extremely useful to invoke a remote builder.

I wanted to use a remote builder because I'm on a slow network. I don't
expect most people to use this, but it's there, (I used it, it worked!)

* make docker-push-support

Put regular docker things back (don't use okteto build or okteto push)

Signed-off-by: Kingdon Barrett <yebyen@gmail.com>